pyproject.toml
[tool.poetry]
name = "cat-detector"
version = "0.1.0"
description = "Prototype cat dection with post-detection action(s)"
authors = ["dax-westerman <daxwesterman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
joblib = "^1.3.2"
numpy = "^1.26.2"
torch = "^2.1.2"
pillow = "^10.1.0"
transformers = "^4.36.2"
opencv-python = "^4.8.1.78"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
